+++ app/src/main/java/com/example/kasapp/data/entity/Transaksi.kt
@@ -9,6 +9,6 @@ data class Transaksi(
     @PrimaryKey(autoGenerate = true)
     val idTransaksi: Int = 0,
     val tglTransaksi: Long = System.currentTimeMillis(), // Simpan sebagai timestamp (Long)
-    val jlhTransaksi: Double, // Total harga
+    val jlhTransaksi: Int, // Total harga
     val jenisPembayaran: String // "QRIS" atau "Tunai"
 )
